ML/requirements.txt
# Unified ML Service Requirements

# Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Machine Learning
scikit-learn==1.3.2
numpy==1.25.2
pandas==2.1.3
joblib==1.3.2

# Computer Vision (parts identification)
Pillow==10.1.0
opencv-python-headless==4.8.1.78

# Database
psycopg2-binary==2.9.9

# Utilities
python-dotenv==1.0.0
pydantic==2.5.0
requests==2.31.0

# Optional: visualization
matplotlib==3.8.2
